authbridge/requirements.txt
python-keycloak>=7.1.1,<8
PyYAML>=6.0.3
